+++ FrameworkSource/UKTestHandler.m
@@ -268,8 +268,8 @@
 
 #pragma mark - Primitive Number Test Assertions
 
-- (void)testInt: (int)a
-        equalTo: (int)b
+- (void)testInt: (long long)a
+        equalTo: (long long)b
          inFile: (const char *)filename
            line: (int)line
 {
@@ -289,8 +289,8 @@
     }
 }
 
-- (void)testInt: (int)a
-     notEqualTo: (int)b
+- (void)testInt: (long long)a
+     notEqualTo: (long long)b
          inFile: (const char *)filename
            line: (int)line
 {
